% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel1}
\alias{bessel1}
\title{Bessel covariance function}
\usage{
bessel1(x, rho)
}
\arguments{
\item{x}{Numeric value}

\item{rho}{First parameter of the distribution}
}
\description{
Bessel covariance function
}
\details{
TBC
}
