% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_lik_grad}
\alias{f_lik_grad}
\title{Calculates the gradient of the full log-likelihood}
\usage{
f_lik_grad(
  B,
  N_dim,
  N_func,
  func_def,
  N_var_func,
  col_id,
  N_par,
  sum_N_par,
  cov_data,
  Z,
  X,
  y,
  u,
  cov_par_fix,
  family,
  link,
  start,
  lower,
  upper,
  tol = 1e-04,
  importance = FALSE
)
}
\arguments{
\item{B}{Integer specifying the number of blocks in the matrix}

\item{N_dim}{Vector of integers, which each value specifying the dimension of each block}

\item{N_func}{Vector of integers specifying the number of functions in the covariance function 
for each block.}

\item{func_def}{Matrix of integers where each column specifies the function definition for each function in each block.}

\item{N_var_func}{Matrix of integers of same size as `func_def` with each column specying the number 
of variables in the argument to each function in each block}

\item{col_id}{3D array (cube) of integers of dimension length(func_def) x max(N_var_func) x B 
where each slice the respective column indexes of `cov_data` for each function in the block}

\item{N_par}{Matrix of integers of same size as `func_def` with each column specifying the number
of parameters in the function in each block}

\item{cov_data}{3D array (cube) holding the data for the covariance matrix where each of the B slices
is the data required for each block}

\item{Z}{Matrix Z of the GLMM}

\item{X}{Matrix X of the GLMM}

\item{y}{Vector of observations}

\item{u}{Matrix of samples of the random effects. Each column is a sample.}

\item{cov_par_fix}{A vector of covariance parameters for importance sampling}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}

\item{start}{Vector of starting values for the optimisation}

\item{lower}{Vector of lower bounds for the model parameters}

\item{upper}{Vector of upper bounds for the model parameters}

\item{trace}{Integer indicating what to report to the console, 0= nothing, 1-3=detailed output}
}
\value{
A vector of the gradient for each parameter
}
\description{
Calculates the gradient of the full log-likelihood
}
