% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/extrafunctions.R
\name{fexp}
\alias{fexp}
\title{Exponential covariance function}
\usage{
fexp(x)

fexp(x)
}
\arguments{
\item{x}{A list with named elements `pars` and `data`. `pars` is a vector with one parameter values, and
`data` is the data `x`}

\item{par1}{First parameter of the distribution}
}
\value{
vector of values of the function
}
\description{
Exponential covariance function

Exponential covariance function
}
\details{
\deqn{\theta_1 exp(-x/\theta_2)}

The function:

\dexp{f(x) = \theta_1*exp(-\theta_2*x)}
}
\examples{
fexp(list(pars = c(1,0.2),data=runif(10)))
}
