\name{glmmrMCML-package}
\alias{glmmrMCML-package}
\alias{glmmrMCML}
\docType{package}
\title{
\packageTitle{glmmrMCML}
}
\description{
\packageDescription{glmmrMCML}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glmmrMCML}
\packageIndices{glmmrMCML}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{glmmrMCML}

Maintainer: \packageMaintainer{glmmrMCML}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
