% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradRobustStep}
\alias{GradRobustStep}
\title{Hill-Climbing algorithm to identify optimal GLMM design}
\usage{
GradRobustStep(
  idx_in,
  n,
  C_list,
  X_list,
  Z_list,
  D_list,
  w_diag,
  max_obs,
  weights,
  exp_cond,
  nfix,
  V0_list,
  any_fix = 0L,
  type = 0L,
  rd_mode = 1L,
  trace = TRUE,
  uncorr = FALSE,
  bayes = FALSE
)
}
\arguments{
\item{idx_in}{Integer vector specifying the indexes of the experimental conditions to start from}

\item{n}{Integer specifying the size of the design to find. For local search, this should be equal to the size of idx_in}

\item{C_list}{List of C vectors for the c-optimal function, see \link{glmmr}[DesignSpace]}

\item{X_list}{List of X matrices}

\item{weights}{Vector specifying the weights of each design}

\item{exp_cond}{Vector specifying the experimental condition index of each observation}

\item{nfix}{Vector listing the experimental condition indexes that are fixed in the design}

\item{any_fix}{Integer. 0 = no experimental conditions are fixed, 1 = some experimental conditions are fixed}

\item{type}{Integer. 0 = local search algorith. 1 = greedy search algorithm.}

\item{rd_mode}{Integer. Robust objective function, 1=weighted average, 2=minimax}

\item{trace}{Logical indicating whether to provide detailed output}

\item{N}{Integer specifying number of experimental conditions in the optimal design}

\item{sig_list}{List of inverse covariance matrices}
}
\value{
A vector of experimental condition indexes in the optimal design
}
\description{
Hill-Climbing algorithm to identify optimal GLMM design
}
