% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_lik_optim}
\alias{l_lik_optim}
\title{Optimises the log-likelihood of the observations conditional on the random effects}
\usage{
l_lik_optim(Z, X, y, u, family, link, start, lower, upper, trace)
}
\arguments{
\item{Z}{Matrix Z of the GLMM}

\item{X}{Matrix X of the GLMM}

\item{y}{Vector of observations}

\item{u}{Matrix of samples of the random effects. Each column is a sample.}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}

\item{start}{Vector of starting values for the optimisation}

\item{lower}{Vector of lower bounds for the model parameters}

\item{upper}{Vector of upper bounds for the model parameters}

\item{trace}{Integer indicating what to report to the console, 0= nothing, 1-3=detailed output}
}
\value{
A vector of parameters that maximise the log likelihood
}
\description{
Optimises the log-likelihood of the observations conditional on the random effects
}
