% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_factorial_approx}
\alias{log_factorial_approx}
\title{Approximation to the log factorial}
\usage{
log_factorial_approx(n)
}
\arguments{
\item{n}{Integer to calculate log(n!)}
}
\value{
A numeric value
}
\description{
Ramanujan's approximation to the log factorial
}
