% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_mv_gaussian_pdf}
\alias{log_mv_gaussian_pdf}
\title{Log multivariate Gaussian probability density funciton}
\usage{
log_mv_gaussian_pdf(u, D, logdetD)
}
\arguments{
\item{u}{Vector of realisations from the distribution}

\item{D}{Inverse covariance matrix}

\item{logdetD}{Log determinant of the covariance matrix}
}
\description{
Log multivariate Gaussian probability density funciton
}
