% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\name{log_mvnd}
\alias{log_mvnd}
\title{Log density of multivariate normal distribution}
\usage{
log_mvnd(s, logdet, invD)
}
\arguments{
\item{s}{vector of observed values}

\item{logdet}{log determinant of the covariance matrix}

\item{invD}{inverse of the covariance matrix}
}
\value{
Matrix
}
\description{
Log density of multivariate normal distribution
}
