% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matern}
\alias{matern}
\title{Matern covariance function}
\usage{
matern(x, rho, nu)
}
\arguments{
\item{x}{Numeric value}

\item{rho}{First parameter of the distribution}

\item{nu}{Second parameter of the distribution}
}
\description{
Matern covariance function
}
\details{
TBC
}
