% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcnr_step}
\alias{mcnr_step}
\title{Newton-Raphson step for the MCMCML algorithm}
\usage{
mcnr_step(y, X, Z, beta, u, family, link)
}
\arguments{
\item{y}{Vector of observations}

\item{X}{Matrix X of the GLMM}

\item{Z}{Matrix Z of the GLMM}

\item{beta}{A vector specifying the current values of the mean function parameters}

\item{u}{Matrix of samples of the random effects. Each column is a sample.}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}
}
\value{
A vector specifying the Newton-Raphson step for the parameters
}
\description{
Newton-Raphson step for the MCMCML algorithm
}
