% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myglm}
\alias{myglm}
\title{Simplified version of fastglm's `fastglm` function}
\usage{
myglm(Xs, ys, weightss, offsets, family)
}
\arguments{
\item{Xs}{input model matrix. Must be a matrix object}

\item{weightss}{an optional vector of 'prior weights' to be used in the fitting process. Should be a numeric vector.}

\item{offsets}{this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be a numeric vector of length equal to the number of cases}

\item{y}{numeric response vector}

\item{A}{\link{stats}[family] object. See `fastglm` for details}
}
\value{
A list with the elements:
* `coefficients` a vector of coefficients
* `se` a vector of the standard errors of the coefficient estimates
* `rank` a scalar denoting the computed rank of the model matrix
* `df.residual` a scalar denoting the degrees of freedom in the model
* `residuals` a vector of residuals
* `s` a numeric scalar - the root mean square for residuals
* `fitted.values` the fitted values
}
\description{
Fast generalized model fitting with a simplified version of fastglm's `fastglm` function
}
\details{
This is a simplified wrapper to the `fastglm` function in the fastglm package, please access that package for more details.
}
