% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{print.glmmr.sim}
\alias{print.glmmr.sim}
\title{Prints a glmmr simuation output}
\usage{
print.glmmr.sim(x, digits = 2, ...)
}
\arguments{
\item{x}{an object of class "`mcml`" as a result of a call to MCML, see \link[glmmr]{Design}}

\item{digits}{Number of digits to print}

\item{...}{Further arguments passed from other methods}
}
\value{
TBC
}
\description{
Print method for class "`glmmr.sim`"
}
\details{
`print.glmmr.sim` calculates multiple statistics summarising the design analysis.

 Simulation diagnostics. TBC
}
