% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qscore_impl}
\alias{qscore_impl}
\title{The quasi-score statistic for a generalised linear mixed model}
\usage{
qscore_impl(resids, tr, xb, invS, family2, weight = TRUE)
}
\arguments{
\item{resids}{A numeric vector of generalised residuals}

\item{tr}{A numeric vector of 1s (treatment group) and -1s (control group)}

\item{xb}{A numeric vector of fitted linear predictors}

\item{invS}{A matrix. If using the weighted statistic then it should be the inverse covariance matrix of the observations}

\item{family2}{A string naming the link function}

\item{weight}{Logical value indicating whether to use the weighted statistic (TRUE) or the unweighted statistic (FALSE)}
}
\value{
A scalar value with the value of the statistic
}
\description{
Generates the quasi-score statistic for a generalised linear mixed model
}
