% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remove_one_many_mat}
\alias{remove_one_many_mat}
\title{Efficiently calculate the inverse of a sub-matrix}
\usage{
remove_one_many_mat(A, i)
}
\arguments{
\item{A}{A square inverse matrix}

\item{i}{Vector of integers specifying the rows/columns to remove from B, see details}
}
\value{
The inverse of a submatrix of B
}
\description{
Efficiently calculate the inverse of a sub-matrix
}
\details{
For a given matrix \eqn{A = B^-1}, this will calculate the inverse of a submatrix of B 
given only matrix A and requiring only vector multiplication. B typically represents a 
covariance matrix for observations 1,...,N and the function is used to calculate the 
inverse covariance matrix for a subset of those observations.
}
\examples{
B <- matrix(runif(16),nrow=4,ncol=4)
diag(B) <- diag(B)+1
A <- solve(B)
remove_one_many_mat(A,1)
#equal to
solve(B[2:4,2:4])
}
