% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sqexp}
\alias{sqexp}
\title{Squared exponential covariance function}
\usage{
sqexp(x, par1, par2)
}
\arguments{
\item{x}{Numeric value}

\item{par1}{First parameter of the distribution}

\item{par2}{Second parameter of the distribution}
}
\description{
Squared exponential covariance function
}
\details{
\deqn{\theta_1 exp(-x^2/\theta_2^2)}
}
