% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{summarize.dfbeta}
\alias{summarize.dfbeta}
\title{Method to summarise DFBETA output}
\usage{
summarize.dfbeta(out)
}
\arguments{
\item{out}{list of mcml model outputs}

\item{n}{Total number of observations in the model}
}
\value{
A list containing the number of observations and proportion of observations
required to change significance, sign, and significant sign from the models.
}
\description{
Method to summarise statistics from the simulation output of `$dfbeta()` from the Design class.
Not generally required by the user.
}
