% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{summarize.errors}
\alias{summarize.errors}
\title{Method to summarise errors}
\usage{
summarize.errors(out, par, true, alpha)
}
\arguments{
\item{out}{list of mcml model outputs}

\item{par}{the index of the parameter to summarise}

\item{true}{true value of the parameter to summarise}

\item{alpha}{the type I error value}
}
\value{
A vector with errors of type 2, M, S1, and S2
}
\description{
Method to summarise errors from the simulation output of `$analysis()` from the Design class.
Not generally required by the user.
}
