% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{summarize.stats}
\alias{summarize.stats}
\title{Method to summarise statistics}
\usage{
summarize.stats(out, par, alpha = 0.05)
}
\arguments{
\item{out}{list of mcml model outputs}

\item{par}{the index of the parameter to summarise}

\item{alpha}{the type I error value}
}
\value{
A list containing two names data frames (`ptot` and `citot`) summarising 
the values of p-statistics from the model fits, and the quantiles of 1-alpha% confidence
interval half-widths, respectively.
}
\description{
Method to summarise statistics from the simulation output of `$analysis()` from the Design class.
Not generally required by the user.
}
