% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aic_mcml}
\alias{aic_mcml}
\title{Calculates the conditional Akaike Information Criterion for the GLMM}
\usage{
aic_mcml(Z, X, y, u, family, link, D_data, beta_par, cov_par)
}
\arguments{
\item{Z}{Matrix Z of the GLMM}

\item{X}{Matrix X of the GLMM}

\item{y}{Vector of observations}

\item{u}{Matrix of samples of the random effects. Each column is a sample.}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}

\item{D_data}{Named list specifying the covariance matrix D. Usually the return from the member function `get_D_data()` of the 
covariance class}

\item{beta_par}{Vector specifying the values of the mean function parameters to estimate the AIC at}

\item{cov_par}{Vector specifying the values of the covariance function parameters to estimate the AIC at}
}
\value{
Estimated conditional AIC
}
\description{
Calculates the conditional Akaike Information Criterion for the GLMM
}
