% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aic_mcml}
\alias{aic_mcml}
\title{Calculates the conditional Akaike Information Criterion for the GLMM}
\usage{
aic_mcml(cov, data, eff_range, Z, X, y, u, family, link, beta_par, cov_par)
}
\arguments{
\item{cov}{An integer matrix with columns of block identifier, dimension of block, function definition, number of variables
in the argument to the funciton, and index of the parameters, respectively. Rows are specific functions of each block.}

\item{data}{Vector of data. Created by flattening the matrices in column-major order of the data used in each block.}

\item{eff_range}{Vector of values with the effective range parameters of the covariance functions, where required.}

\item{Z}{Matrix Z of the GLMM}

\item{X}{Matrix X of the GLMM}

\item{y}{Vector of observations}

\item{u}{Matrix of samples of the random effects. Each column is a sample.}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}

\item{beta_par}{Vector specifying the values of the mean function parameters to estimate the AIC at}

\item{cov_par}{Vector specifying the values of the covariance function parameters to estimate the AIC at}
}
\value{
Estimated conditional AIC
}
\description{
Calculates the conditional Akaike Information Criterion for the GLMM
}
