% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_u_samples.R
\name{gen_u_samples}
\alias{gen_u_samples}
\title{Generate samples of random effects using MCMC}
\usage{
gen_u_samples(y, X, Z, L, beta, family, sigma = 1, warmup_iter = 100, m = 100)
}
\arguments{
\item{y}{Vector of outcomes}

\item{X}{Matrix of covariates}

\item{Z}{Matrix, the design matrix of the random effects}

\item{L}{Matrix, the Cholesky decomposition of the covariance matrix of the random effects}

\item{beta}{Vector of mean function parameters}

\item{family}{A family function, e.g. gaussian()}

\item{sigma}{Numeric, the scale parameter of the distribution}

\item{warmup_iter}{Numeric, the number of warmup iterations}

\item{m}{Numeric, the number of sampling iterations}
}
\value{
A matrix in which each column is a sample of the random effects
}
\description{
Generate samples of random effects using MCMC
}
\details{
Calls Stan through `cmdstanr` to generate `m` samples of the random effects from
a GLMM conditional on fixed values of the model parameters. To make use of 
parallelisation, the model is parameterised in terms of the Cholesky decomposition
of the covariance matrix of the random effects. Only a single chain is run to generate 
the samples.
}
