\name{glmmrMCML-package}
\alias{glmmrMCML-package}
\alias{glmmrMCML}
\docType{package}
\title{
\packageTitle{glmmrMCML}
}
\description{
\packageDescription{glmmrMCML}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glmmrMCML}
\packageIndices{glmmrMCML}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{glmmrMCML}

Maintainer: \packageMaintainer{glmmrMCML}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
Add links
}
\examples{
## Not run
}
