% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcml_hess}
\alias{mcml_hess}
\title{Generate Hessian matrix of GLMM}
\usage{
mcml_hess(D_data, Q, Z, X, y, u, family, link, start, tol = 1e-05, trace = 0L)
}
\arguments{
\item{D_data}{Named list specifying the covariance matrix D. Usually the return from the member function `get_D_data()` of the 
covariance class}

\item{Q}{Integer specifying the number of covariance parameters}

\item{Z}{Matrix Z of the GLMM}

\item{X}{Matrix X of the GLMM}

\item{y}{Vector of observations}

\item{u}{Matrix of samples of the random effects. Each column is a sample.}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}

\item{start}{Vector of starting values for the optimisation}

\item{tol}{The tolerance of the numerical differentiation routine}

\item{trace}{Integer indicating what to report to the console, 0= nothing, 1-3=detailed output}

\item{cov_par_fix}{A vector of covariance parameters for importance sampling}
}
\value{
The estimated Hessian matrix
}
\description{
Generate Hessian matrix of GLMM using numerical differentiation
}
\details{
Member function `$get_D_data()` of the covariance class will provide the necessary objects to specify the covariance matrix
}
