% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcml_hess}
\alias{mcml_hess}
\title{Generate Hessian matrix of GLMM}
\usage{
mcml_hess(
  cov,
  data,
  eff_range,
  Z,
  X,
  y,
  u,
  family,
  link,
  start,
  tol = 1e-05,
  trace = 0L
)
}
\arguments{
\item{cov}{An integer matrix with columns of block identifier, dimension of block, function definition, number of variables
in the argument to the funciton, and index of the parameters, respectively. Rows are specific functions of each block.}

\item{data}{Vector of data. Created by flattening the matrices in column-major order of the data used in each block.}

\item{eff_range}{Vector of values with the effective range parameters of the covariance functions, where required.}

\item{Z}{Matrix Z of the GLMM}

\item{X}{Matrix X of the GLMM}

\item{y}{Vector of observations}

\item{u}{Matrix of samples of the random effects. Each column is a sample.}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}

\item{start}{Vector of starting values for the optimisation}

\item{tol}{The tolerance of the numerical differentiation routine}

\item{trace}{Integer indicating what to report to the console, 0= nothing, 1-3=detailed output}
}
\value{
The estimated Hessian matrix
}
\description{
Generate Hessian matrix of GLMM using numerical differentiation
}
\details{
Member function `$get_D_data()` of the covariance class will provide the necessary objects to specify the covariance matrix
}
