% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcml_la_nr}
\alias{mcml_la_nr}
\title{Maximum Likelihood with Laplace Approximation and Newton-Raphson}
\usage{
mcml_la_nr(
  cov,
  data,
  eff_range,
  Z,
  X,
  y,
  family,
  link,
  start,
  usehess = FALSE,
  tol = 0.001,
  verbose = TRUE,
  trace = 0L,
  maxiter = 10L
)
}
\arguments{
\item{cov}{An integer matrix with columns of block identifier, dimension of block, function definition, number of variables
in the argument to the funciton, and index of the parameters, respectively. Rows are specific functions of each block.}

\item{data}{Vector of data. Created by flattening the matrices in column-major order of the data used in each block.}

\item{eff_range}{Vector of values with the effective range parameters of the covariance functions, where required.}

\item{Z}{Matrix Z of the GLMM}

\item{X}{Matrix X of the GLMM}

\item{y}{Vector of observations}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}

\item{start}{Vector of starting values for the optimisation}

\item{tol}{Value of the tolerance. The algorithm termninates if differences in values of parameters between
iterations are all less than this value.}

\item{verbose}{Logical indicating whether to provide output to the console}

\item{trace}{Integer indicating what to report to the console, 0= nothing, 1-3=ever more detailed output}
}
\description{
Maximum Likelihood with Laplace Approximation and Newton-Raphson
}
