% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcml_optim_sparse}
\alias{mcml_optim_sparse}
\title{Likelihood maximisation for the GLMM using sparse matrix methods}
\usage{
mcml_optim_sparse(
  cov,
  data,
  eff_range,
  Ap,
  Ai,
  Z,
  X,
  y,
  u,
  family,
  link,
  start,
  trace,
  mcnr = FALSE
)
}
\arguments{
\item{cov}{An integer matrix with columns of block identifier, dimension of block, function definition, number of variables
in the argument to the funciton, and index of the parameters, respectively. Rows are specific functions of each block.}

\item{data}{Vector of data. Created by flattening the matrices in column-major order of the data used in each block.}

\item{eff_range}{Vector of values with the effective range parameters of the covariance functions, where required.}

\item{Ap}{Integer vector of pointers, one for each column, specifying the initial (zero-based) index of elements in the column. Slot `p`
of a matrix of a class defined in \link[Matrix]{sparseMatrix}}

\item{Ai}{Integer vector specifying the row indices of the non-zero elements of the matrix. Slot `i`
of a matrix of a class defined in \link[Matrix]{sparseMatrix}}

\item{Z}{Matrix Z of the GLMM}

\item{X}{Matrix X of the GLMM}

\item{y}{Vector of observations}

\item{u}{Matrix of samples of the random effects. Each column is a sample.}

\item{family}{Character specifying the family}

\item{link}{Character specifying the link function}

\item{start}{Vector of starting values for the optimisation}

\item{trace}{Integer indicating what to report to the console, 0= nothing, 1-3=detailed output}

\item{mcnr}{Logical indicating whether to use Newton-Raphson (TRUE) or Expectation Maximisation (FALSE)}
}
\value{
A vector of the parameters that maximise the simulated likelihood
}
\description{
Given model data and random effects samples `u`, this function will run the MCML steps to generate new estimates of 
mean function and covariance parameters
}
\details{
Member function `$get_D_data()` of the covariance class will provide the necessary objects to specify the covariance matrix

Likelihood maximisation for the GLMM
}
