% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ModelExtMCML.R
\name{mcnr_family}
\alias{mcnr_family}
\title{Returns the file name and type for MCNR function}
\usage{
mcnr_family(family)
}
\arguments{
\item{family}{family object}
}
\value{
list with filename and type
}
\description{
Returns the file name and type for MCNR function
}
