% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvn_ll}
\alias{mvn_ll}
\title{Multivariate normal log likelihood}
\usage{
mvn_ll(cov, data, eff_range, gamma, u)
}
\arguments{
\item{cov}{An integer matrix with columns of block identifier, dimension of block, function definition, number of variables
in the argument to the funciton, and index of the parameters, respectively. Rows are specific functions of each block.}

\item{data}{Vector of data. Created by flattening the matrices in column-major order of the data used in each block.}

\item{eff_range}{Vector of values with the effective range parameters of the covariance functions, where required.}

\item{gamma}{Vector specifying the values of the covariance function parameters}

\item{u}{Matrix (or vector) of observed values}
}
\value{
Scalar value
}
\description{
Calculates the log likelihood of the multivariate normal distribution using `glmmr` covariance representation
}
